requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
frozenlist==1.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
pyee==8.2.2
python-dotenv==0.19.2
slackclient==2.9.3
slackeventsapi==3.0.1
Werkzeug==2.0.2
yarl==1.7.2
